pyproject.toml
[tool.poetry]
name = "cipher_shh2145_2"
version = "0.1.0"
description = "This package contains the cipher function to encode and decode secret messages."
authors = ["Shauna Han"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
myst-nb = "^0.17.1"
sphinx-autoapi = "^2.0.0"
sphinx-rtd-theme = "^1.1.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
